+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/customer/CustomerDraft.java
@@ -265,6 +265,14 @@ public interface CustomerDraft extends com.commercetools.api.models.Customizable
     public AuthenticationMode getAuthenticationMode();
 
     /**
+     *  <p>Customer Groups to assign to the Customer.</p>
+     * @return customerGroupAssignments
+     */
+    @Valid
+    @JsonProperty("customerGroupAssignments")
+    public List<CustomerGroupAssignmentDraft> getCustomerGroupAssignments();
+
+    /**
      *  <p>User-defined unique identifier for the Customer. The <code>key</code> field is preferred over <code>customerNumber</code> as it is mutable and provides more flexibility.</p>
      * @param key value to be set
      */
@@ -498,6 +506,21 @@ public interface CustomerDraft extends com.commercetools.api.models.Customizable
     public void setAuthenticationMode(final AuthenticationMode authenticationMode);
 
     /**
+     *  <p>Customer Groups to assign to the Customer.</p>
+     * @param customerGroupAssignments values to be set
+     */
+
+    @JsonIgnore
+    public void setCustomerGroupAssignments(final CustomerGroupAssignmentDraft... customerGroupAssignments);
+
+    /**
+     *  <p>Customer Groups to assign to the Customer.</p>
+     * @param customerGroupAssignments values to be set
+     */
+
+    public void setCustomerGroupAssignments(final List<CustomerGroupAssignmentDraft> customerGroupAssignments);
+
+    /**
      * factory method
      * @return instance of CustomerDraft
      */
@@ -539,6 +562,7 @@ public interface CustomerDraft extends com.commercetools.api.models.Customizable
         instance.setSalutation(template.getSalutation());
         instance.setStores(template.getStores());
         instance.setAuthenticationMode(template.getAuthenticationMode());
+        instance.setCustomerGroupAssignments(template.getCustomerGroupAssignments());
         return instance;
     }
 
@@ -592,6 +616,11 @@ public interface CustomerDraft extends com.commercetools.api.models.Customizable
                         .collect(Collectors.toList()))
                 .orElse(null));
         instance.setAuthenticationMode(template.getAuthenticationMode());
+        instance.setCustomerGroupAssignments(Optional.ofNullable(template.getCustomerGroupAssignments())
+                .map(t -> t.stream()
+                        .map(com.commercetools.api.models.customer.CustomerGroupAssignmentDraft::deepCopy)
+                        .collect(Collectors.toList()))
+                .orElse(null));
         return instance;
     }