+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/customer/CustomerDraftImpl.java
@@ -77,6 +77,8 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
 
     private com.commercetools.api.models.customer.AuthenticationMode authenticationMode;
 
+    private java.util.List<com.commercetools.api.models.customer.CustomerGroupAssignmentDraft> customerGroupAssignments;
+
     /**
      * create instance with all properties
      */
@@ -101,7 +103,8 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
             @JsonProperty("custom") final com.commercetools.api.models.type.CustomFieldsDraft custom,
             @JsonProperty("locale") final String locale, @JsonProperty("salutation") final String salutation,
             @JsonProperty("stores") final java.util.List<com.commercetools.api.models.store.StoreResourceIdentifier> stores,
-            @JsonProperty("authenticationMode") final com.commercetools.api.models.customer.AuthenticationMode authenticationMode) {
+            @JsonProperty("authenticationMode") final com.commercetools.api.models.customer.AuthenticationMode authenticationMode,
+            @JsonProperty("customerGroupAssignments") final java.util.List<com.commercetools.api.models.customer.CustomerGroupAssignmentDraft> customerGroupAssignments) {
         this.key = key;
         this.customerNumber = customerNumber;
         this.externalId = externalId;
@@ -129,6 +132,7 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
         this.salutation = salutation;
         this.stores = stores;
         this.authenticationMode = authenticationMode;
+        this.customerGroupAssignments = customerGroupAssignments;
     }
 
     /**
@@ -361,6 +365,14 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
         return this.authenticationMode;
     }
 
+    /**
+     *  <p>Customer Groups to assign to the Customer.</p>
+     */
+
+    public java.util.List<com.commercetools.api.models.customer.CustomerGroupAssignmentDraft> getCustomerGroupAssignments() {
+        return this.customerGroupAssignments;
+    }
+
     public void setKey(final String key) {
         this.key = key;
     }
@@ -488,6 +500,16 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
         this.authenticationMode = authenticationMode;
     }
 
+    public void setCustomerGroupAssignments(
+            final com.commercetools.api.models.customer.CustomerGroupAssignmentDraft... customerGroupAssignments) {
+        this.customerGroupAssignments = new ArrayList<>(Arrays.asList(customerGroupAssignments));
+    }
+
+    public void setCustomerGroupAssignments(
+            final java.util.List<com.commercetools.api.models.customer.CustomerGroupAssignmentDraft> customerGroupAssignments) {
+        this.customerGroupAssignments = customerGroupAssignments;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o)
@@ -525,6 +547,7 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
                 .append(salutation, that.salutation)
                 .append(stores, that.stores)
                 .append(authenticationMode, that.authenticationMode)
+                .append(customerGroupAssignments, that.customerGroupAssignments)
                 .append(key, that.key)
                 .append(customerNumber, that.customerNumber)
                 .append(externalId, that.externalId)
@@ -552,6 +575,7 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
                 .append(salutation, that.salutation)
                 .append(stores, that.stores)
                 .append(authenticationMode, that.authenticationMode)
+                .append(customerGroupAssignments, that.customerGroupAssignments)
                 .isEquals();
     }
 
@@ -584,6 +608,7 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
                 .append(salutation)
                 .append(stores)
                 .append(authenticationMode)
+                .append(customerGroupAssignments)
                 .toHashCode();
     }
 
@@ -616,6 +641,7 @@ public class CustomerDraftImpl implements CustomerDraft, ModelBase {
                 .append("salutation", salutation)
                 .append("stores", stores)
                 .append("authenticationMode", authenticationMode)
+                .append("customerGroupAssignments", customerGroupAssignments)
                 .build();
     }