+++ commercetools/commercetools-sdk-java-api/src/test/java-generated/com/commercetools/api/models/message/CustomerGroupAssignmentAddedMessageTest.java
@@ -1,0 +1,39 @@
+
+package com.commercetools.api.models.message;
+
+import com.tngtech.junit.dataprovider.DataProvider;
+import com.tngtech.junit.dataprovider.DataProviderExtension;
+import com.tngtech.junit.dataprovider.UseDataProvider;
+import com.tngtech.junit.dataprovider.UseDataProviderExtension;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestTemplate;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+@ExtendWith(UseDataProviderExtension.class)
+@ExtendWith(DataProviderExtension.class)
+public class CustomerGroupAssignmentAddedMessageTest {
+
+    @TestTemplate
+    @UseDataProvider("objectBuilder")
+    public void buildUnchecked(CustomerGroupAssignmentAddedMessageBuilder builder) {
+        CustomerGroupAssignmentAddedMessage customerGroupAssignmentAddedMessage = builder.buildUnchecked();
+        Assertions.assertThat(customerGroupAssignmentAddedMessage)
+                .isInstanceOf(CustomerGroupAssignmentAddedMessage.class);
+    }
+
+    @DataProvider
+    public static Object[][] objectBuilder() {
+        return new Object[][] { new Object[] { CustomerGroupAssignmentAddedMessage.builder()
+                .customerGroupAssignment(new com.commercetools.api.models.customer.CustomerGroupAssignmentImpl()) } };
+    }
+
+    @Test
+    public void customerGroupAssignment() {
+        CustomerGroupAssignmentAddedMessage value = CustomerGroupAssignmentAddedMessage.of();
+        value.setCustomerGroupAssignment(new com.commercetools.api.models.customer.CustomerGroupAssignmentImpl());
+        Assertions.assertThat(value.getCustomerGroupAssignment())
+                .isEqualTo(new com.commercetools.api.models.customer.CustomerGroupAssignmentImpl());
+    }
+}