+++ changes.md
@@ -9,6 +9,8 @@
 - added property `approvalRuleMode` to type `CompanyDraft`
 - added property `approvalRuleMode` to type `Division`
 - added property `approvalRuleMode` to type `DivisionDraft`
+- added property `customerGroupAssignments` to type `Customer`
+- added property `customerGroupAssignments` to type `CustomerDraft`
 - added property `metaTitle` to type `ProductTailoringCreatedMessage`
 - added property `metaDescription` to type `ProductTailoringCreatedMessage`
 - added property `metaKeywords` to type `ProductTailoringCreatedMessage`
@@ -115,12 +117,23 @@
 - added type `CustomerSearchIndexingStatusResponse`
 - added type `CustomerSearchRequest`
 - added type `CustomerSearchResult`
+- added type `CustomerGroupAssignment`
+- added type `CustomerGroupAssignmentDraft`
+- added type `CustomerAddCustomerGroupAssignmentAction`
+- added type `CustomerRemoveCustomerGroupAssignmentAction`
+- added type `CustomerSetCustomerGroupAssignmentsAction`
 - added type `BusinessUnitApprovalRuleModeChangedMessage`
+- added type `CustomerGroupAssignmentAddedMessage`
+- added type `CustomerGroupAssignmentRemovedMessage`
+- added type `CustomerGroupAssignmentsSetMessage`
 - added type `ProductTailoringImageAddedMessage`
 - added type `ProductTailoringImagesSetMessage`
 - added type `ProductVariantTailoringAddedMessage`
 - added type `ProductVariantTailoringRemovedMessage`
 - added type `BusinessUnitApprovalRuleModeChangedMessagePayload`
+- added type `CustomerGroupAssignmentAddedMessagePayload`
+- added type `CustomerGroupAssignmentRemovedMessagePayload`
+- added type `CustomerGroupAssignmentsSetMessagePayload`
 - added type `ProductTailoringImageAddedMessagePayload`
 - added type `ProductTailoringImagesSetMessagePayload`
 - added type `ProductVariantTailoringAddedMessagePayload`